services/jokes/requirements.txt
Flask==2.3.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.5
pycryptodome==3.18.0
Requests==2.31.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
gunicorn